% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentral.R
\name{BetaDistribution}
\alias{BetaDistribution}
\alias{pbeta_boost}
\alias{qbeta_boost}
\title{The non-central Beta distribution}
\usage{
pbeta_boost(q, a, b, ncp = 0, lower.tail = TRUE)

qbeta_boost(p, a, b, ncp = 0, lower.tail = TRUE)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{a, b}{shape parameters, positive numbers}

\item{ncp}{non-centrality parameter, a non-negative number}

\item{lower.tail}{Boolean, whether to use the complementary probabilities}

\item{p}{numeric vector of probabilities}
}
\value{
A numeric vector of probabilities (\code{pbeta_boost}) or a 
  numeric vector of quantiles (\code{qbeta_boost}).
}
\description{
Distribution function and quantile function for the non-central 
  Beta distribution.
}
