% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentral.R
\name{find_chisq_nu}
\alias{find_chisq_nu}
\title{Find degrees of freedom parameter}
\usage{
find_chisq_nu(ncp, q, p)
}
\arguments{
\item{ncp}{non-centrality parameter, a non-negative number}

\item{q}{a quantile}

\item{p}{probability corresponding to the quantile \code{q}}
}
\value{
The degrees of freedom parameter of the non-central Chi-squared 
  distribution with cumulative probability \code{p} at the quantile \code{q}.
}
\description{
Find the degrees of freedom parameter of a non-central 
  Chi-squared distribution given a quantile and its corresponding probability.
}
\examples{
nu <- find_chisq_nu(10, 3, 0.1)
pchisq(3, df = nu, ncp = 10) # should be 0.1
}
