% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentral.R
\name{FisherDistribution}
\alias{FisherDistribution}
\alias{pf_boost}
\alias{qf_boost}
\title{The non-central Fisher distribution}
\usage{
pf_boost(q, nu1, nu2, ncp = 0, lower.tail = TRUE)

qf_boost(p, nu1, nu2, ncp = 0, lower.tail = TRUE)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{nu1}{numerator degrees of freedom parameter, a positive number}

\item{nu2}{denominator degrees of freedom parameter, a positive number}

\item{ncp}{non-centrality parameter, a non-negative number}

\item{lower.tail}{Boolean, whether to use the complementary probabilities}

\item{p}{numeric vector of probabilities}
}
\value{
A numeric vector of probabilities (\code{pf_boost}) or a 
  numeric vector of quantiles (\code{qf_boost}).
}
\description{
Distribution function and quantile function for the non-central 
  Fisher distribution.
}
