% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentral.R
\name{StudentDistribution}
\alias{StudentDistribution}
\alias{pt_boost}
\alias{qt_boost}
\title{The non-central Student distribution}
\usage{
pt_boost(q, nu, ncp = 0, lower.tail = TRUE)

qt_boost(p, nu, ncp = 0, lower.tail = TRUE)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{nu}{degrees of freedom parameter, a positive number}

\item{ncp}{non-centrality parameter}

\item{lower.tail}{Boolean, whether to use the complementary probabilities}

\item{p}{numeric vector of probabilities}
}
\value{
A numeric vector of probabilities (\code{pt_boost}) or a 
  numeric vector of quantiles (\code{qt_boost}).
}
\description{
Distribution function and quantile function for the non-central 
  Student distribution.
}
\examples{
pt_boost(100, nu = 2, ncp = 100)
}
